separate pkgaction into 'pkgaction' and 'serviceaction'
ignored IntellJ IDEA files
