+++ .gitignore
@@ -1,2 +1,4 @@
 docker-machine*
 *.log
+*.iml
+.idea/
