+++ libmachine/host.go
@@ -15,6 +15,7 @@ import (
 	"github.com/docker/machine/libmachine/engine"
 	"github.com/docker/machine/libmachine/provision"
 	"github.com/docker/machine/libmachine/provision/pkgaction"
+	"github.com/docker/machine/libmachine/provision/serviceaction"
 	"github.com/docker/machine/libmachine/swarm"
 	"github.com/docker/machine/log"
 	"github.com/docker/machine/ssh"
@@ -244,7 +245,7 @@ func (h *Host) Upgrade() error {
 		return err
 	}
 
-	if err := provisioner.Service("docker", pkgaction.Restart); err != nil {
+	if err := provisioner.Service("docker", serviceaction.Restart); err != nil {
 		return err
 	}
 	return nil
