+++ libmachine/provision/pkgaction/pkg_action.go
@@ -1,0 +1,23 @@
+package pkgaction
+
+type PackageAction int
+
+const (
+	Install PackageAction = iota
+	Remove
+	Upgrade
+)
+
+var packageActions = []string{
+	"install",
+	"remove",
+	"upgrade",
+}
+
+func (s PackageAction) String() string {
+	if int(s) >= 0 && int(s) < len(packageActions) {
+		return packageActions[s]
+	}
+
+	return ""
+}
