+++ libmachine/provision/rancheros.go
@@ -10,6 +10,7 @@ import (
 	"github.com/docker/machine/libmachine/auth"
 	"github.com/docker/machine/libmachine/engine"
 	"github.com/docker/machine/libmachine/provision/pkgaction"
+	"github.com/docker/machine/libmachine/provision/serviceaction"
 	"github.com/docker/machine/libmachine/swarm"
 	"github.com/docker/machine/log"
 	"github.com/docker/machine/state"
@@ -49,7 +50,7 @@ type RancherProvisioner struct {
 	GenericProvisioner
 }
 
-func (provisioner *RancherProvisioner) Service(name string, action pkgaction.ServiceAction) error {
+func (provisioner *RancherProvisioner) Service(name string, action serviceaction.ServiceAction) error {
 	command := fmt.Sprintf("sudo system-docker %s %s", action.String(), name)
 
 	if _, err := provisioner.SSHCommand(command); err != nil {
