+++ libmachine/provision/serviceaction/service_action.go
@@ -1,4 +1,4 @@
-package pkgaction
+package serviceaction
 
 type ServiceAction int
 
@@ -23,28 +23,6 @@ var serviceActions = []string{
 func (s ServiceAction) String() string {
 	if int(s) >= 0 && int(s) < len(serviceActions) {
 		return serviceActions[s]
-	}
-
-	return ""
-}
-
-type PackageAction int
-
-const (
-	Install PackageAction = iota
-	Remove
-	Upgrade
-)
-
-var packageActions = []string{
-	"install",
-	"remove",
-	"upgrade",
-}
-
-func (s PackageAction) String() string {
-	if int(s) >= 0 && int(s) < len(packageActions) {
-		return packageActions[s]
 	}
 
 	return ""
